Fix bottom sheet
Fixed HintView.
CategotyView monolith is broken into structured components
Fixed colour scheme.
Fixed component design
